added swagger config, category service and controller

+++ src/main/java/com/filmflix/vodservice/db/repositories/TagRepository.java
@@ -3,5 +3,5 @@ package com.filmflix.vodservice.db.repositories;
 import com.filmflix.vodservice.db.entities.Tag;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TagRepository extends JpaRepository<Tag, Long> {
+public interface TagRepository extends JpaRepository<Tag, Integer> {
 }
